requirements.txt
graphene_django
django_storages
boto3
pillow
